Move command utils over to logic folder
